host judgement

+++ project/src/JDBCExample.java
@@ -31,7 +31,7 @@ public class JDBCExample {
     private static String username;
 
     // some strings
-    private static final String is_owner_prompt = "1: Personal Information, 2: Manage My Listings, 3: Listing Availability, 4: Check Booking, 5: rating, 9: logout";
+    private static final String is_owner_prompt = "1: Personal Information, 2: Manage My Listings, 3: Listing Availability, 4: Check Booking, 5: Rating, 9: logout";
     private static final String is_renter_prompt = "1: Personal Information, 2: Manage My Booking, 5: Rating, 9: logout";
     private static final String a_line = "--------------------------------------------------------------------------------";
     private static final String half_line = "-----------------------------------";
@@ -559,6 +559,47 @@ public class JDBCExample {
                             }
                         } else if (input.equals("5")) {
                             System.out.println("Rating and Comment");
+
+							System.out.println("Press the Corresponding number to continue");
+                            System.out.println("1: Show how I was being rated as a host, 2: Show what I've rated, 3: Rate a renter, 4: Delete a rating, 5: Go Back.");
+                            String booking_decide = sc.nextLine();
+                            if (booking_decide.equals("1")) {
+                                print_header("Show how I was being rated");
+                                if (!show_host_s_judgement(username)) {
+                                    print_header("No renter has rated you so far.");
+                                }
+                            } else if (booking_decide.equals("2")){
+								print_header("Show what I've rated");
+								if (!show_host_s_rate(username)) {
+                                    print_header("You haven't rate anyone yet.");
+                                }
+							} else if (booking_decide.equals("3")) {
+                                print_header("Rate a renter");
+                                System.out.println("Please input the renter's username that you want to rate");
+                                String renter_username = validate(sc);
+                                System.out.println("Please input how much you like this renter from 1-5");
+                                String likes = validate_int(sc, 1, 5);
+								System.out.println("Please make any comment (100 words limit):");
+                                String words = validate(sc);
+
+								create_host_rate(renter_username, username, likes, words);
+                            } else if (booking_decide.equals("4")) {
+                                // cannot change a book
+                                print_header("Delete a rating");
+                                System.out.println("Input the JID that you want to cancel");
+                                if (!show_host_s_rate(username)) {
+                                    print_header("you haven't rate anyone");
+                                }
+                                String jid = validate_int(sc, 1, 9999);
+                                if(!cancel_judgement_host(jid, username)){
+                                    print_error("Cannot cancel the rate");
+                                }
+                            } else if (booking_decide.equals("5")) {
+                                //go back
+                            }
+
+
+
                         } else if (input.equals("8")) {
                             System.out.println("Terminate the program.");
                             break label_whole;
@@ -716,8 +757,8 @@ public class JDBCExample {
                                     print_header("you don't have any booking");
                                 }
                                 String jid = validate_int(sc, 1, 9999);
-                                if(!cancel_judgement(jid)){
-                                    print_error("Cannot cancel book");
+                                if(!cancel_judgement_renter(jid, username)){
+                                    print_error("Cannot cancel the rate");
                                 }
                             } else if (booking_decide.equals("5")) {
                                 //go back
@@ -1378,12 +1419,45 @@ public class JDBCExample {
             return false;
         }
     }
-    public static boolean cancel_judgement(String jid) throws SQLException {
+	public static boolean cancel_judgement_host(String jid, String hostname) throws SQLException {
         try {
-            String sql = String.format("DELETE FROM Judgement where JID = \"%s\";", jid);
-            System.out.print(sql.concat("\n"));
-            stmt.executeUpdate(sql);
-            return true;
+
+			String sql = String.format("SELECT * FROM Judgement where host_username = '%s' and direction = 1;", hostname);
+			ResultSet rs = stmt.executeQuery(sql);
+			while(rs.next()){
+					String JID = rs.getString("JID");
+					if(JID.equals(jid)){
+						String sql2 = String.format("DELETE FROM Judgement where JID = \"%s\";", jid);
+						System.out.print(sql.concat("\n"));
+						stmt.executeUpdate(sql2);
+						return true;
+					}
+				}
+			    print_header("Invalid JID: please select from what you have rated");
+				return false;
+            
+        } catch (SQLException e) {
+            e.printStackTrace();
+            return false;
+        }
+    }
+	public static boolean cancel_judgement_renter(String jid, String rentername) throws SQLException {
+        try {
+
+			String sql = String.format("SELECT * FROM Judgement where renter_username = '%s' and direction = 0;", rentername);
+			ResultSet rs = stmt.executeQuery(sql);
+			while(rs.next()){
+					String JID = rs.getString("JID");
+					if(JID.equals(jid)){
+						String sql2 = String.format("DELETE FROM Judgement where JID = \"%s\";", jid);
+						System.out.print(sql.concat("\n"));
+						stmt.executeUpdate(sql2);
+						return true;
+					}
+				}
+			    print_header("Invalid JID: please select from what you have rated");
+				return false;
+            
         } catch (SQLException e) {
             e.printStackTrace();
             return false;
@@ -1474,6 +1548,37 @@ public class JDBCExample {
 			return false;
 		}
 	}
+	public static boolean create_host_rate(String rentername, String hostname, String rate, String text) throws SQLException{
+		try{
+
+				String sql = String.format("SELECT Book.username FROM Book Join Owns where Book.lid = Owns.lid and cancellation = 0 and Owns.username = '%s';", hostname);
+				ResultSet rs = stmt.executeQuery(sql);
+				while(rs.next()){
+					String renter = rs.getString("username");
+					if(renter.equals(rentername)){
+						String sql3 = String.format("SELECT * FROM Judgement where renter_username = '%s' and host_username = '%s' and direction = 1;", rentername, hostname);
+						ResultSet rs3 = stmt.executeQuery(sql3);
+						if(rs3.next()){
+							System.out.println("Since you have already rated this renter, we delete the previous one and create new rating.");
+							String JID =rs3.getString("JID");
+							String sql4 = String.format("DELETE FROM Judgement where JID = '%s';", JID);
+							stmt.executeUpdate(sql4);
+						}
+						String sql2 = String.format("INSERT INTO Judgement (words, host_username, renter_username, direction, likes) VALUES ('%s', '%s', '%s', '%s', '%s');", text, hostname,rentername,"1",rate);
+						print_header("Rating added.");
+						stmt.executeUpdate(sql2);
+						return true;
+					}
+					
+				}
+				print_header("You can't judge this renter since he/she hasn't book any from you.");
+				return false;
+	
+		}catch(SQLException e){
+			System.err.println("You can't judge this renter since he/she hasn't book any from you.");
+			return false;
+		}
+	}
    /*show how the renter is being judged*/
 	public static boolean show_renter_s_judgement(String rentername) throws SQLException{
 		try{
@@ -1501,6 +1606,33 @@ public class JDBCExample {
 			return false;
 		}
 	}
+	/*show how the host is being judged*/
+	public static boolean show_host_s_judgement(String hostname) throws SQLException{
+		try{
+				boolean result = false;
+				String sql = String.format("SELECT * FROM Judgement where host_username = '%s';", hostname);
+				ResultSet rs = stmt.executeQuery(sql);
+				while(rs.next()){
+					String JID = rs.getString("JID");
+					String likes = rs.getString("likes");
+					String words = rs.getString("words");
+					String direction = rs.getString("direction");
+					String renter_username = rs.getString("renter_username");
+
+					if(direction.equals("0")){
+						String sql2 = String.format("JID = '%s', Renter: '%s', Likes = '%s', Comment: '%s';", JID, renter_username, likes, words);
+						System.out.print(sql2.concat("\n"));
+						//stmt.executeUpdate(sql2);
+						result = true;
+					}
+				}
+				return result;
+	
+		}catch(SQLException e){
+			System.err.println("Something went wrong with judgement showing.");
+			return false;
+		}
+	}
 	/*Show what renter have rated*/
 	public static boolean show_renter_s_rate(String rentername) throws SQLException{
 		try{
@@ -1516,6 +1648,33 @@ public class JDBCExample {
 
 					if(direction.equals("0")){
 						String sql2 = String.format("JID = '%s', Host: '%s', Likes = '%s', Comment: '%s';", JID, host_username,likes,words);
+						System.out.print(sql2.concat("\n"));
+						//stmt.executeUpdate(sql2);
+						result = true;
+					}
+				}
+				return result;
+	
+		}catch(SQLException e){
+			System.err.println("Something went wrong with judgement showing.");
+			return false;
+		}
+	}
+	/*Show what host have rated*/
+	public static boolean show_host_s_rate(String hostname) throws SQLException{
+		try{
+				boolean result = false;
+				String sql = String.format("SELECT * FROM Judgement where host_username = '%s';", hostname);
+				ResultSet rs = stmt.executeQuery(sql);
+				while(rs.next()){
+					String JID = rs.getString("JID");
+					String likes = rs.getString("likes");
+					String words = rs.getString("words");
+					String direction = rs.getString("direction");
+					String renter_username = rs.getString("renter_username");
+
+					if(direction.equals("1")){
+						String sql2 = String.format("JID = '%s', Renter: '%s', Likes = '%s', Comment: '%s';", JID, renter_username,likes,words);
 						System.out.print(sql2.concat("\n"));
 						//stmt.executeUpdate(sql2);
 						result = true;
